# application/services/custom_services/feature/src/models/entitymodels/entity.ts
export interface entity 
{
   created_date_at: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date_at: { type: Date },
   name: String,
   age: Number
}